NIFI-7922: Added support for GET in ListenHTTP for health check
NIFI-7922: Fixes based on review comments

ListenHTTP: changed if(!sslRequired) to if(sslRequired) so that the positive case comes first.
HealthCheckServlet: response body for GET contains "OK".
ContentAcknowledgmentServlet: super.doDelete() is called when a DELETE should be rejected because of port mismatch.

NIFI-7922: Refactored, based on review comments

NIFI-7922: Fixed a checkstyle violation (organized the imports)

This closes #4603.

Signed-off-by: Peter Turcsanyi <turcsanyi@apache.org>